part 4.21

+++ controllers/blogs.js
@@ -75,6 +75,23 @@ blogsRouter.post('/', async (request,response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    //console.log("user id", blog.user.toString())
+    //console.log("token id", decodedToken.id)
+
+    if(blog.user.toString() !== decodedToken.id) {
+      return response.status(401).json({ error: 'unauthorized resource' })  
+    }
+
+
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
